Move leb128 module to util
